+++ backend/src/test/java/site/coduo/acceptance/TimerAcceptanceTest.java
@@ -61,7 +61,7 @@ class TimerAcceptanceTest extends AcceptanceFixture {
         return response.accessCode();
     }
 
-    private static void timerStart(final String accessCode) {
+    public static void timerStart(final String accessCode) {
         createConnect(accessCode);
         RestAssured
                 .given()
@@ -162,87 +162,5 @@ class TimerAcceptanceTest extends AcceptanceFixture {
 
                 .then()
                 .statusCode(204);
-    }
-
-    @Test
-    @DisplayName("타이머 비활성화 요칭 시,로그인 쿠키가 없으면 400을 반환한다.")
-    void return_400_status_code_without_login_cookie() {
-        // given
-        final PairRoomCreateRequest pairRoom = new PairRoomCreateRequest("fram", "lemone", 10000L,
-                10000L, "https://missionUrl.xxx", PairRoomStatus.IN_PROGRESS.name());
-        final String userId = "userId";
-        final String token = jwtProvider.sign(userId);
-        saveMember(userId);
-        final String accessCode = createPairRoom(pairRoom, token);
-
-        // when & then
-        RestAssured
-                .given()
-                .when()
-                .pathParam("accessCode", accessCode)
-                .patch("/api/{accessCode}/timer/disable")
-
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    @DisplayName("타이머 기능을 비활성화는 SSE커넥션이 있는 상태에서만 가능하다.")
-    void disable_pair_room_timer() {
-        // given
-        final PairRoomCreateRequest pairRoom = new PairRoomCreateRequest("fram", "lemone", 10000L,
-                10000L, "https://missionUrl.xxx", PairRoomStatus.IN_PROGRESS.name());
-        final String userId = "userId";
-        final String token = jwtProvider.sign(userId);
-        saveMember(userId);
-        final String accessCode = createPairRoom(pairRoom, token);
-        createConnect(accessCode);
-        TimerAcceptanceTest.timerStart(accessCode);
-
-        // when & then
-        RestAssured
-                .given()
-                .cookie("coduo_whoami", token)
-                .when()
-                .pathParam("accessCode", accessCode)
-                .patch("/api/{accessCode}/timer/disable")
-
-                .then().log().all()
-                .statusCode(200);
-    }
-
-    private void saveMember(final String userId) {
-        final Member member = Member.builder()
-                .username("test user")
-                .userId(userId)
-                .loginId(FakeGithubApiClient.LOGIN_ID)
-                .accessToken(FakeGithubOAuthClient.ACCESS_TOKEN.getCredential())
-                .profileImage(FakeGithubApiClient.PROFILE_IMAGE)
-                .build();
-
-        memberRepository.save(member);
-    }
-
-    @Test
-    @DisplayName("자신의 방이 아닌 방에 타이머를 비활성 요칭 시,  403 상태코드를 반환한다.")
-    void return_403_status_code_when_timer_disable_request_to_does_not_exist_pair_room() {
-        // given
-        final PairRoomCreateRequest pairRoom = new PairRoomCreateRequest("fram", "lemone", 10000L,
-                10000L, "https://missionUrl.xxx", PairRoomStatus.IN_PROGRESS.name());
-        final String userId = "userId";
-        final String token = jwtProvider.sign(userId);
-        saveMember(userId);
-
-        // when & then
-        RestAssured
-                .given()
-                .cookie("coduo_whoami", token)
-
-                .when()
-                .pathParam("accessCode", "does not exist")
-                .patch("/api/{accessCode}/timer/disable")
-
-                .then().log().all()
-                .statusCode(403);
     }
 }
