+++ backend/src/test/java/site/coduo/sync/service/SchedulerServiceTest.java
@@ -2,7 +2,6 @@ package site.coduo.sync.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
 import java.util.concurrent.TimeUnit;
@@ -17,11 +16,6 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 
 import site.coduo.fake.FakeScheduledFuture;
 import site.coduo.member.exception.AuthorizationException;
-import site.coduo.pairroom.domain.MissionUrl;
-import site.coduo.pairroom.domain.Pair;
-import site.coduo.pairroom.domain.PairName;
-import site.coduo.pairroom.domain.PairRoom;
-import site.coduo.pairroom.domain.PairRoomStatus;
 import site.coduo.pairroom.domain.accesscode.AccessCode;
 import site.coduo.pairroom.repository.PairRoomEntity;
 import site.coduo.pairroom.service.PairRoomService;
@@ -127,15 +121,12 @@ class SchedulerServiceTest {
     void remove_timestamp_in_timestamp_registry() {
         // given
         final String key = "some-access-code";
-        final String userId = "fram";
         eventStreamsRegistry.register(key);
         schedulerRegistry.register(key, new FakeScheduledFuture());
         timestampRegistry.register(key, new Timer(new AccessCode(key), 10000L, 10000L));
-        when(pairRoomService.isParticipant(anyString(), anyString()))
-                .thenReturn(true);
 
         // when
-        schedulerService.detach(key, "accessCode");
+        schedulerService.detach(key);
 
         // then
         assertThat(timestampRegistry.has(key)).isFalse();
@@ -150,11 +141,9 @@ class SchedulerServiceTest {
         eventStreamsRegistry.register(key);
         schedulerRegistry.register(key, new FakeScheduledFuture());
         timestampRegistry.register(key, new Timer(new AccessCode(key), 10000L, 10000L));
-        when(pairRoomService.isParticipant(anyString(), anyString()))
-                .thenReturn(true);
 
         // when
-        schedulerService.detach(key, "accessToken");
+        schedulerService.detach(key);
 
         // then
         assertThat(schedulerRegistry.has(key)).isFalse();
@@ -168,11 +157,9 @@ class SchedulerServiceTest {
         eventStreamsRegistry.register(key);
         schedulerRegistry.register(key, new FakeScheduledFuture());
         timestampRegistry.register(key, new Timer(new AccessCode(key), 10000L, 10000L));
-        when(pairRoomService.isParticipant(anyString(), anyString()))
-                .thenReturn(true);
 
         // when
-        schedulerService.detach(key, "accessToken");
+        schedulerService.detach(key);
 
         // then
         assertThat(eventStreamsRegistry.hasNoStreams(key)).isTrue();
@@ -186,21 +173,10 @@ class SchedulerServiceTest {
         eventStreamsRegistry.register(key);
         schedulerRegistry.register(key, new FakeScheduledFuture());
         timestampRegistry.register(key, new Timer(new AccessCode(key), 10000L, 10000L));
-        when(pairRoomService.isParticipant(anyString(), anyString()))
-                .thenReturn(false);
 
         // when & then
-        assertThatThrownBy(() -> schedulerService.detach(key, "accessToken"))
+        assertThatThrownBy(() -> schedulerService.detach(key))
                 .isInstanceOf(AuthorizationException.class);
     }
 
-    private PairRoomEntity createPairRoom(final String code, final PairRoomStatus pairRoomStatus) {
-        final AccessCode accessCode = new AccessCode(code);
-        return PairRoomEntity.from(
-                new PairRoom(pairRoomStatus,
-                        new Pair(new PairName("레디"), new PairName("레모네")),
-                        new MissionUrl("https://missionUrl.xxx"),
-                        accessCode
-                ));
-    }
 }
